First post to github
Converted all the parts to step and uploaded. Im still working on various parts, nothing is set in stone yet. Do not make this, because im not sure whats going to be changing or not in the future. Right now its just me playing around designing a printer that no one has a want, need or use for.